GUI is available!
Updates to main. Extra modules are in the /classes folder for organization purposes. Note that the code is a bit of a mess right now, but at least its working (mostly). Right now it only works with compilers that use only the name and input file. No additional switches are available yet.